+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### `jsonschema-generator-bom`
+#### Added
+- Introduce BOM (Bill of Materials) artifact, as option for importing Generator + Modules with matching versions
+
 ### `jsonschema-maven-plugin`
 #### Added
-- Declared thread-safety, to avoid warnings at runtime
+- Declare thread-safety, to avoid warnings at runtime
 
 ## [4.22.0] - 2022-01-10
 ### `jsonschema-generator`